Refactor `ErrorType` with key renamed
