+++ src/controllers/Definition/DefinitionGetController.ts
@@ -5,8 +5,7 @@ import * as DefinitionSelectDAO from '../../dao/Definition/DefinitionSelectDAO';
 import * as PosSelectDAO from '../../dao/Pos/PosSelectDAO';
 import * as UsageSelectDAO from '../../dao/Usage/UsageSelectDAO';
 import * as OriginSelectDAO from '../../dao/Origin/OriginSelectDAO';
-import MarmoymError from '../../models/MarmoymError';
-import ErrorType from '../../models/ErrorType';
+import MarmoymError from '@models/MarmoymError';
 import { transaction } from '../../database/databaseUtils';
 import { DefinitionResponse } from '../../routes/ResponseTypes';
 import * as RequestTypes from '../../routes/RequestTypes';
