+++ src/controllers/User/UserDeleteController.ts
@@ -3,7 +3,7 @@ import * as bcrypt from 'bcrypt';
 import db from '../../database';
 import * as UserDeleteDAO from '../../dao/User/UserDeleteDAO';
 import { transaction } from '../../database/databaseUtils';
-import MarmoymError from "../../models/MarmoymError";
+import MarmoymError from "@models/MarmoymError";
 import ErrorType from '../../models/ErrorType';
 
 export function deleteUser(req) {
@@ -13,7 +13,7 @@ export function deleteUser(req) {
     if (userDeleted == 1) {
       return 'UserDeleteSuccess';
     } else {
-      throw new MarmoymError(ErrorType.User.USER_DELETE_FAIL);
+      throw new MarmoymError(ErrorType.USER.USER_DELETE_FAIL);
     }
     
   });
