+++ src/models/ErrorType/AuthErrorType.ts
