+++ src/models/ErrorType/TermErrorType.ts
